Add FFT to Scope.
Currently, Scope can show voltage, current, power, and a few other neat things.
It can sometimes guess the dominant frequency of the voltage.

However, one of the more useful features of real oscilloscopes that is still
missing is the ability to show the full Fourier transform of a signal.

This patch adds the "Show FFT" option to Scope.
Enabling this option will cause Scope to use the new FFT class to compute the
FFT of the voltage and paint the magnitude of the FFT as a bright red line
behind the voltage/current/power.