CirSim.java: lu_factor() - clean up and simplify singular matrix check
lu_factor() would create and update a local temporary array
scaleFactors[] with scaling factors for the matrix, but it never used
the values in the array for anything; it was simply wasted work. This
code has been in the simulator at least since 1.3g from November 2008.

This commit eliminates this array and all references to it.

It also restructures lu_factor()'s initial test for a singular matrix
with a row of all zeroes. Since we're no longer building scale factors,
instead of scanning each row for the element with the largest
magnitude, we simply scan it for any nonzero elements. As soon as we
find such an element, we stop scanning that row and skip on to the next
one. Only if all the elements in the row are zero do we return false
to signal a singular matrix.

I've tested this change a fair amount. While it may have a small
performance benefit for nonlinear circuits, which must call lu_factor()
for every timestep, this is probably insignificant. Mainly it just
simplifies the logic a little.