TappedTransformerElm - user settings for coupling and trap. vs. Euler
TransformerElm has user settings for the transformer's coupling
coefficient and for whether its differential equations are solved
using the trapezoidal method or the backward-Euler method.
TappedTranformerElm currently lacks these settings; this change adds
them to that element as well. TappedTransformerElm's default coupling
coefficient is 0.99, the same value as is used already; although this
differs from TransformerElm's default value, it maintains
compatibility with existing circuits using TappedTransformerElm.

I've tested this change with all the sample and test circuits that
currently use TappedTransformerElm; they appear to work properly, and
changes in the settings work as expected.

I did not include the secondary-polarity setting from TransformerElm,
since that is more complex than changing the other settings and thus
seemed to be better handled in a separate commit.

My biggest motivation for doing this was to allow a
TappedTransformerElm to use the backward Euler method for integration.
Although the backward Euler method technically produces a less accurate
solution (one with a greater numeric error) than the trapezoidal method
for a given timestep size, it has one major advantage over the
trapezoidal method in certain situations.

The trapezoidal method may produce oscillating solutions in circuits
which change rapidly compared to the timestep, such as switching
circuits. This is an inherent property of the method. The backward
Euler method doesn't behave this way, so for certain circuits you can
get smoother, more qualitatively realistic behavior using that method,
even though the numeric error in the values is technically greater.