Add lineOver to ring counter's reset pin
Circuitjs's ring counter has an active-low reset pin and is cleared with a low reset signal (this is distinct from the active-high reset pins of devices such as the CD4026BE). Unless I'm mistaken, I believe this should be marked with an overline.
I spent a solid 10 minutes trying to figure out why the ring counter was not working - the lack of marking suggested the reset pin is active-high, not active-low.
Related but not part of this commit: It would be nice to have an edit option to change whether or not the reset pin is active-high or active-low.