WebSocket: properly handle initialization events
We need to register two events: one for the iframe to be loaded, which
will then give us a fresh reference to the contentWindow object. On that
contentWindow we can then attach the CircuitJS event which will fire
once the simulator has fully loaded. This way, no racy behavior is
exhibited and we also don't need to poll the simulator periodically.